Move utility to utils/dataframe.py
