+++ baybe/objectives/single.py
@@ -10,9 +10,8 @@ from typing_extensions import override
 
 from baybe.objectives.base import Objective
 from baybe.targets.base import Target
-from baybe.utils.dataframe import pretty_print_df
+from baybe.utils.dataframe import get_transform_objects, pretty_print_df
 from baybe.utils.plotting import to_string
-from baybe.utils.validation import get_transform_objects
 
 
 @define(frozen=True, slots=False)
