+++ baybe/utils/dataframe.py
@@ -4,11 +4,7 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Collection, Iterable, Sequence
-from typing import (
-    TYPE_CHECKING,
-    Literal,
-    overload,
-)
+from typing import TYPE_CHECKING, Literal, TypeVar, overload
 
 import numpy as np
 import pandas as pd
@@ -21,7 +17,10 @@ from baybe.utils.numerical import DTypeFloatNumpy
 if TYPE_CHECKING:
     from torch import Tensor
 
-    from baybe.parameters import Parameter
+    from baybe.parameters.base import Parameter
+    from baybe.targets.base import Target
+
+    _T = TypeVar("_T", bound=Parameter | Target)
 
 # Logging
 _logger = logging.getLogger(__name__)
@@ -503,3 +502,51 @@ def pretty_print_df(
         )
         str_df = str(str_df)
     return str_df
+
+
+def get_transform_objects(
+    objects: Sequence[_T],
+    df: pd.DataFrame,
+    allow_missing: bool,
+    allow_extra: bool,
+) -> list[_T]:
+    """Extract the objects relevant for transforming a given dataframe.
+
+    The passed object are assumed to have corresponding columns in the given dataframe,
+    identified through their name attribute. The function returns the subset of objects
+    that have a corresponding column in the dataframe and thus provide the necessary
+    information for transforming the dataframe.
+
+    Args:
+        objects: A collection of objects to be considered for transformation (provided
+            they have a match in the given dataframe).
+        df: The dataframe to be searched for corresponding columns.
+        allow_missing: Flag controlling if objects are allowed to have no corresponding
+            columns in the dataframe.
+        allow_extra: Flag controlling if the dataframe is allowed to have columns
+            that have no corresponding objects.
+
+    Raises:
+        ValueError: If the given objects and dataframe are not compatible
+            under the specified values for the Boolean flags.
+
+    Returns:
+        The (subset of) objects that need to be considered for the transformation.
+    """
+    names = [p.name for p in objects]
+
+    if (not allow_missing) and (missing := set(names) - set(df)):  # type: ignore[arg-type]
+        raise ValueError(
+            f"The object(s) named {missing} cannot be matched against "
+            f"the provided dataframe. If you want to transform a subset of "
+            f"columns, explicitly set `allow_missing=True`."
+        )
+
+    if (not allow_extra) and (extra := set(df) - set(names)):
+        raise ValueError(
+            f"The provided dataframe column(s) {extra} cannot be matched against"
+            f"the given objects. If you want to transform a dataframe "
+            f"with additional columns, explicitly set `allow_extra=True'."
+        )
+
+    return [p for p in objects if p.name in df]
